pyproject.toml
[project]
name = "rnmon"
version = "0.2.0"
description = "Reticulum Monitoring Daemon"
readme = "README.md"
authors = [
    { name = "lbatalha" }
]
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "rns>=0.9.6",
]

[project.scripts]
rnmon = "rnmon.rnmon:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
